proper ui setup
