+++ src/Anyshift.ini
@@ -30,7 +30,7 @@ neutral detection = True
 ; time the shifter must stay in neutral to engage it
 neutral delay = 0.5
 ; true if clutch is needed to change gears
-require clutch = False
+require clutch = True
 ; percent of the clutch that needs to be pressed to allow gear change
 clutch bitepoint = 75
 ; true if the game uses a separated button for reverse. gran turismo or nascar racing for example
