+++ src/Global.py
@@ -1,0 +1,15 @@
+#########################################################################################################
+# Global variables between modules
+#
+# There is no warranty for the program, to the extent permitted by applicable law. Except 
+# when otherwise stated in writing the copyright holders and/or other parties provide the
+# program "as is" without warranty of any kind, either expressed or implied, including, but
+# not limited to, the implied warranties of merchantability and fitness for a particular purpose.  
+# The entire risk as to the quality and performance of the program is with you.  
+# Should the program prove defective, you assume the cost of all necessary servicing, 
+# repair or correction.
+#
+# 2023 Menkaura Soft
+#########################################################################################################
+
+done = False
